// server/src/models/original.metadata.model.ts
export default interface OriginalGalleryMetadata {
    gid: number
    token: string
    title: string
    title_jpn: string
    category: string
    thumb: string
    uploader: string
    posted: number
    filecount: number
    filesize: number
    expunged: boolean
    rating: number
    torrentcount: number
    tags: string[]
}